refactor : 기능 우선 구조로 인한 auth의 data 하위 폴더로 이동
